podcast-frontend-api/requirements.txt
flask[async]>=2.0.0,<2.1.0
werkzeug>=2.0.0,<2.1.0
azure-servicebus==7.8.3
python-dotenv==1.0.0
gunicorn==20.1.0
pydantic>=2.0.0
psutil>=5.9.0
azure-mgmt-monitor>=6.0.0
azure-identity>=1.12.0
aiohttp>=3.8.0